+++ extension/src/panel/zodiac-routes/index.ts
@@ -1,2 +1,4 @@
 export * from './routeHooks'
+export { useSelectedRouteId } from './SelectedRouteContext'
 export { useZodiacRoute } from './useZodiacRoute'
+export { ProvideZodiacRoutes, useZodiacRoutes } from './ZodiacRouteContext'
